NoteApp/backend/requirements.txt
fastapi[standard]
uvicorn
sqlite3-binary ; platform_system != "Windows"
PyJWT
passlib[bcrypt]
python-multipart
